Rename widgets/{selectbox->select}.ts
