+++ bokehjs/src/lib/models/widgets/index.ts
@@ -26,7 +26,7 @@ export {PasswordInput}       from "./password_input"
 export {PreText}             from "./pretext"
 export {RadioButtonGroup}    from "./radio_button_group"
 export {RadioGroup}          from "./radio_group"
-export {Select}              from "./selectbox"
+export {Select}              from "./select"
 export {Spinner}             from "./spinner"
 export {Switch}              from "./switch"
 export {TextAreaInput}       from "./textarea_input"
